fix memory leak on null icon
if the icon cannot be loaded (e.g. it is an empty string), then the
pixmap is null, and thus rotatePixmap will return a null pointer which
is assigned back to the same variable.

easy solution: if the pixmap is null, delete it instead, and rotate it
only if not